Introduced picking up the ball
Ball is dropped on ragdolling (hopefully when tackled or tripped)

Collision for ball is still on so wipes you out if you try to slide etc - to be fixed.

Animation work is maybe needed, or an IK solution

Needs to be set up for both hands too.

Introduced bug:

Title text does not fade